requirements.txt
langgraph
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
pypdf
python-dotenv
requests